+++ android/app/src/main/java/poke/rogue/helper/presentation/poketmon2/PokemonStatAdapter.kt
@@ -5,6 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import poke.rogue.helper.databinding.ItemStatBinding
+import poke.rogue.helper.presentation.poketmon2.detail.Stat
 import poke.rogue.helper.presentation.util.view.ItemDiffCallback
 
 class PokemonStatAdapter() :
